// packages/shared/src/generated/EntitlementSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntitlementKey } from "./EntitlementKey";
import type { FeatureValueSource } from "./FeatureValueSource";

export type EntitlementSnapshot = { key: EntitlementKey, enabled: boolean, limit: bigint | null, source: FeatureValueSource, };