// packages/shared/src/generated/FeatureAccessSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntitlementSnapshot } from "./EntitlementSnapshot";
import type { FeatureFlagSnapshot } from "./FeatureFlagSnapshot";
import type { PlanId } from "./PlanId";

export type FeatureAccessSnapshot = { planId: PlanId, flags: Array<FeatureFlagSnapshot>, entitlements: Array<EntitlementSnapshot>, };