// packages/shared/src/generated/FeatureFlagSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FeatureFlagDefinition } from "./FeatureFlagDefinition";
import type { FeatureFlagKey } from "./FeatureFlagKey";
import type { FeatureValueSource } from "./FeatureValueSource";

export type FeatureFlagSnapshot = { key: FeatureFlagKey, enabled: boolean, source: FeatureValueSource, definition: FeatureFlagDefinition, overrideEnabled: boolean | null, };